#23 refactor and renaming
