+++ .github/PULL_REQUEST_TEMPLATE/PULL_REQUEST_TEMPLATE.md
@@ -8,7 +8,7 @@ Used issue # (fill in your issue)
 
 - [ ] New Component (non-breaking change which adds functionality to the website)
 - [ ] New Feature (non-breaking change which adds a new css component)
-- [ ] Bug fix (non-breaking change which fixes an bug)
+- [ ] Bug fix (non-breaking change which fixes a bug)
 - [ ] Breaking change (fix or feature that would cause existing functionality to not work as expected)
 
 # Checklist:
@@ -21,7 +21,7 @@ Used issue # (fill in your issue)
 - [ ] My changes generate no new warnings or errors inside the console.
 - [ ] It's easy to see what issue was used for these changes.
 - [ ] I have only used module css classes, with exception for some part where you can use the index.scss.
-- [ ] I have read the [`CODE OF CONDUCT`](CODE_OF_CONDUCT.md) & the [`CONTRIBUTING`](CONTRIBUTING.md).
+- [ ] I have read the [`CODE OF CONDUCT`](../CODE_OF_CONDUCT.md) & the [`CONTRIBUTING`](../CONTRIBUTING.md).
 
 ## Components specific
