+++ .github/SECURITY.md
@@ -3,8 +3,8 @@
 ## Supported Versions
 
 | Version | Supported          |
-| ------- | ------------------ |
-| latest   | :white_check_mark:|
+|---------|--------------------|
+| latest  | :white_check_mark: |
 
 ## Reporting a Vulnerability
