+++ config/tsconfig.node.json
@@ -18,5 +18,7 @@
     "noUnusedParameters": true,
     "noFallthroughCasesInSwitch": true
   },
-  "include": ["vite.config.ts"]
+  "include": [
+    "../vite.config.ts"
+  ]
 }
