+++ tsconfig.json
@@ -1,7 +1,7 @@
 {
   "files": [],
   "references": [
-    { "path": "./tsconfig.app.json" },
-    { "path": "./tsconfig.node.json" }
+    { "path": "./config/tsconfig.app.json"},
+    { "path": "./config/tsconfig.node.json"}
   ]
 }
